Added documentation to the esi module and ApiErrors
