+++ esi.gemspec
@@ -28,4 +28,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'rake', '~> 10.0'
   spec.add_development_dependency 'rubocop', '~> 0.52'
   spec.add_development_dependency 'shoulda', '~> 3.5'
+  spec.add_development_dependency 'yard', '~> 0.9'
 end
